Move sun conditions to the sun integration
